requirements.txt
wand
pandas
Pillow
